Добавить хуки по умолчанию
